requirements.txt

# Required - Core packages
numpy>=2.0.2
scipy>=1.13.1
pandas>=2.2.3

# Required - Machine learning (required)
scikit-learn>=1.5.2
scikit-image>=0.25.2
umap-learn>=0.5.7

# Required - Deep learning (required)
torch>=2.6.0
torchvision>=0.21.0
deepTools>=0.1.9

# Required - Image I/O and manipulation
opencv-python>=4.10.0
imageio>=2.37.0
Pillow>=10.4.0
colour>=0.1.5

# Required - Image visualization
matplotlib>=3.9.2
matplotlib-base>=3.9.4
seaborn>=0.13.2
graphviz>=0.20.3
plotly>=6.0.0

# Required - Tabular results
openpyxl>=3.1.5

# Required - Symbolic mathematics
sympy>=1.13.3

# Required - For GUI Interface
gradio>=5.16.1
gradio_client>=1.7.0
